refactored IUnitOfWork and Repository Interfaces into the Applicatoin Layer from the Domain Layer
