+++ Core/Application/LibSystem.Application/Common/Interfaces/IUnitOfWork.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LibSystem.Domain.Common
+namespace LibSystem.Application.Common.Interfaces
 {
     public interface IUnitOfWork
     {
