+++ Core/Application/LibSystem.Application/DependencyInjection.cs
@@ -19,7 +19,7 @@ namespace LibSystem.Application
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
 
             // Register FluentValidation validators from this assembly
-            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly()); // Ensure FluentValidation is installed and referenced
+            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly()); 
 
             // Register MediatR with handlers from this assembly
             services.AddMediatR(cfg =>
@@ -27,8 +27,7 @@ namespace LibSystem.Application
                 cfg.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly());
             });
 
-            // Register pipeline behaviors in order of execution
-            // Order matters: Validation -> Logging -> Performance -> Unhandled Exception
+            // Registering Pipeline behaviour in Order matters: Validation -> Logging -> Performance -> Unhandled Exception
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(UnhandledExceptionBehaviour<,>));
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehavior<,>));
             services.AddTransient(typeof(IPipelineBehavior<,>), typeof(LoggingBehaviour<,>));
